+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## [5.10.1-bb.0] - 2022-05-04
+### Changed
+- Update GitLab to app version 14.10.1 and helm chart to version v5.10.1
+- Removed sidecar injection from several jobs
+### Added
+- Maintenance documentation for how to upgrade and test
+
 ## [5.9.2-bb.2] - 2022-04-26
 ### Added
 - Added Oscal Component